type of output matrix changed

+++ tasks/seq/kruglov_a_components_marking/func_tests/main.cpp
@@ -15,8 +15,8 @@ TEST(kruglov_a_components_marking_seq_functional, test_functional) {
                              1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1,
                              1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0,
                              1, 1, 1, 0, 0, 1, 1, 1, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1};
-  std::vector<uint8_t> out(h * w, 0);
-  std::vector<uint8_t> comp = {1, 1, 0, 0, 2, 2, 2, 0, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 0, 0, 1, 1, 0, 0, 0,
+  std::vector<uint32_t> out(h * w, 0);
+  std::vector<uint32_t> comp = {1, 1, 0, 0, 2, 2, 2, 0, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 0, 0, 1, 1, 0, 0, 0,
                                0, 0, 0, 0, 0, 0, 1, 0, 0, 3, 0, 3, 0, 0, 0, 0, 0, 0, 0, 3, 3, 3, 0, 0, 0,
                                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 4, 0, 0, 4, 0, 0, 0, 0, 0, 0, 4, 4, 4, 4,
                                0, 0, 0, 5, 5, 0, 0, 0, 0, 4, 0, 0, 6, 0, 0, 0, 0, 0, 0, 0, 6, 6, 6, 0, 0};
@@ -47,8 +47,8 @@ TEST(kruglov_a_components_marking_seq_functional, test_non_square) {
   std::vector<uint32_t> size = {h, w};
   std::vector<uint8_t> in = {0, 0, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1,
                              1, 1, 1, 1, 1, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 1, 1, 1};
-  std::vector<uint8_t> out(h * w, 0);
-  std::vector<uint8_t> comp = {1, 1, 0, 0, 2, 2, 2, 0, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 0, 0, 1, 1, 0, 0, 0,
+  std::vector<uint32_t> out(h * w, 0);
+  std::vector<uint32_t> comp = {1, 1, 0, 0, 2, 2, 2, 0, 0, 0, 1, 0, 0, 0, 0, 0, 2, 0, 0, 0, 1, 1, 0, 0, 0,
                                0, 0, 0, 0, 0, 0, 1, 0, 0, 3, 0, 3, 0, 0, 0, 0, 0, 0, 0, 3, 3, 3, 0, 0, 0};
 
   // Create TaskData
@@ -75,8 +75,8 @@ TEST(kruglov_a_components_marking_seq_functional, test_all_ones) {
   uint32_t w = 10;
   std::vector<uint32_t> size = {h, w};
   std::vector<uint8_t> in(h * w, 1);
-  std::vector<uint8_t> out(h * w, 0);
-  std::vector<uint8_t> comp(h * w, 0);
+  std::vector<uint32_t> out(h * w, 0);
+  std::vector<uint32_t> comp(h * w, 0);
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -102,8 +102,8 @@ TEST(kruglov_a_components_marking_seq_functional, test_all_zeros) {
   uint32_t w = 10;
   std::vector<uint32_t> size = {h, w};
   std::vector<uint8_t> in(h * w, 0);
-  std::vector<uint8_t> out(h * w, 0);
-  std::vector<uint8_t> comp(h * w, 1);
+  std::vector<uint32_t> out(h * w, 0);
+  std::vector<uint32_t> comp(h * w, 1);
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
@@ -129,12 +129,12 @@ TEST(kruglov_a_components_marking_seq_functional, test_grid) {
   uint32_t w = 10;
   std::vector<uint32_t> size = {h, w};
   std::vector<uint8_t> in(h * w);
-  std::vector<uint8_t> comp(h * w);
+  std::vector<uint32_t> comp(h * w);
   for (uint32_t i = 0; i < h * w; ++i) {
     in[i] = (i / 10 + i % 10) % 2;
     comp[i] = ((i / 10 + i % 10) % 2 == 0) ? i / 2 + 1 : 0;
   }
-  std::vector<uint8_t> out(h * w, 0);
+  std::vector<uint32_t> out(h * w, 0);
 
   // Create TaskData
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
